Finish making Toychain modular
The Forest.v module is minimal, i.e. it only exports lemmas and
definitions that are actually used elsewhere in the proof.

The modules in Systems/ are NOT minimal and are structured in a way
that means the same code is copy/pasted -- the root cause is the fact
that Coq's kernel doesn't understand that Parameters can be
instantiated by Inductives.  TODO: Find a better way of doing this.