Gossip protocol is a world of pain :-(
Representing (IPv4, port) pairs as ints theoretically "works", but is
practically unusable because of performance. I was expecting
arithmethic operations to be impractical, but I didn't have any use
for them anyway.

However, I did not anticipate that you can't do _any_ operation on
these numbers in Gallina. Not even comparison. One immediate
implication of this is that you can't even compare them, which you
need fo ConnectMsg, AddrMsg and GetDataMsg.

What this means is that as of now, receiving a ConnectMsg, AddrMsg or
GetDataMsg will, for all practical purposes, crash your node.